+++ tests/testthat/test-eplet_registry.R
@@ -93,14 +93,18 @@ test_that("right alleles are returned", {
   expect_equal(lookup_alleles(df_eplets, "23L"), list(`23L` = "DQB1*04:01"))
   expect_equal(
     lookup_alleles(df_eplets, "4Q", allele_set = "all"),
-    list(`4Q` = c(
-      "DRB1*01:144", "DRB1*07:01", "DRB1*07:03", "DRB1*07:04", "DRB1*07:07",
-      "DRB1*07:09", "DRB1*07:121", "DRB1*07:136", "DRB1*07:139", "DRB1*07:151",
-      "DRB1*07:152", "DRB1*07:153", "DRB1*09:01", "DRB1*09:20", "DRB1*09:21",
-      "DRB1*09:31", "DRB1*09:32", "DRB1*09:57", "DRB4*01:01", "DRB4*01:02",
-      "DRB4*01:03", "DRB4*01:72", "DRB4*01:151", "DRB4*01:152", "DRB4*01:155",
-      "DRB4*01:156", "DRB4*01:168", "DRB4*01:173", "DRB4*01:176"
-    ))
+    list(
+      `4Q` = c(
+        "DRB1*01:144", "DRB1*07:01", "DRB1*07:03", "DRB1*07:04", "DRB1*07:07",
+        "DRB1*07:09", "DRB1*07:34", "DRB1*07:121", "DRB1*07:136", "DRB1*07:139",
+        "DRB1*07:151", "DRB1*07:152", "DRB1*07:153", "DRB1*07:159",
+        "DRB1*09:01", "DRB1*09:20", "DRB1*09:21", "DRB1*09:31", "DRB1*09:32",
+        "DRB1*09:57", "DRB1*09:58", "DRB1*09:59", "DRB4*01:01", "DRB4*01:02",
+        "DRB4*01:03", "DRB4*01:07", "DRB4*01:151", "DRB4*01:152",
+        "DRB4*01:155", "DRB4*01:156", "DRB4*01:168", "DRB4*01:173",
+        "DRB4*01:176"
+      )
+    )
   )
 })
 
@@ -233,9 +237,7 @@ test_that("low cardinality character columns contain expected values", {
 
 test_that("table has no empty eplets/alleles", {
   expect_true(sum(is.na(df_eplets$name) | df_eplets$name == "") == 0)
-  warning("3P (DQ) has no Luminex alleles in v. 2024-03-15 of eplet registry;
-          this test was forced to succeed")
-  succeed(sum(is.na(df_eplets$alleles) | df_eplets$alleles == "") == 0)
+  expect_true(sum(is.na(df_eplets$alleles) | df_eplets$alleles == "") == 0)
 })
 
 test_that("table has no duplicate eplets", {
@@ -309,8 +311,8 @@ test_that("a few randomly selected cells have same value as on the website", {
       "alleles"
     ),
     c(
-      "DQB1*06:103", "DQB1*06:205", "DQB1*06:243", "DQB1*06:359",
-      "DQB1*06:382", "DQB1*06:415", "DQB1*06:472"
+      "DQB1*06:01", "DQB1*06:103", "DQB1*06:205", "DQB1*06:243", "DQB1*06:359",
+      "DQB1*06:382", "DQB1*06:415", "DQB1*06:472", "DQB1*06:482"
     )
   )
 })